+++ Package.swift
@@ -1,7 +1,7 @@
 // swift-tools-version:5.9
 import PackageDescription
 
-let version = "2.1.4"
+let version = "2.1.5"
 
 let package = Package(
     name: "osx-bench",